renamed Upload component to Encode
