+++ pdf-2-base64/src/app/components/encode.js
