+++ pdf-2-base64/src/app/page.js
@@ -1,7 +1,7 @@
 'use client'
 import Image from "next/image";
 import styles from "./page.module.css";
-import Upload from "./components/upload";
+import Encode from "./components/encode";
 import { useState } from "react";
 
 export default function Home() {
@@ -16,7 +16,7 @@ export default function Home() {
     <div className={styles.page}>
       <main className={styles.main}>
         <h3>PDF Encoder For IQ x Soon</h3>
-        <Upload prop={SetFileString}/>
+        <Encode prop={SetFileString}/>
         <div>Base64 String <button onClick={copyToClipboard}>COPY</button>:</div>
         <div className={styles.div}>{FileString}</div>
       </main>
